+++ torch_geometric/contrib/degree_quant/environment.yml
@@ -1,0 +1,49 @@
+name: graphs
+channels:
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=main
+  - _openmp_mutex=5.1=1_gnu
+  - ca-certificates=2023.01.10=h06a4308_0
+  - certifi=2022.12.7=py38h06a4308_0
+  - ld_impl_linux-64=2.38=h1181459_1
+  - libffi=3.4.2=h6a678d5_6
+  - libgcc-ng=11.2.0=h1234567_1
+  - libgomp=11.2.0=h1234567_1
+  - libstdcxx-ng=11.2.0=h1234567_1
+  - ncurses=6.4=h6a678d5_0
+  - openssl=1.1.1t=h7f8727e_0
+  - pip=22.3.1=py38h06a4308_0
+  - python=3.8.16=h7a1cb2a_2
+  - readline=8.2=h5eee18b_0
+  - setuptools=65.6.3=py38h06a4308_0
+  - sqlite=3.40.1=h5082296_0
+  - tk=8.6.12=h1ccaba5_0
+  - wheel=0.38.4=py38h06a4308_0
+  - xz=5.2.10=h5eee18b_1
+  - zlib=1.2.13=h5eee18b_0
+  - pip:
+      - charset-normalizer==3.0.1
+      - idna==3.4
+      - jinja2==3.1.2
+      - joblib==1.2.0
+      - markupsafe==2.1.2
+      - numpy==1.24.2
+      - nvidia-cublas-cu11==11.10.3.66
+      - nvidia-cuda-nvrtc-cu11==11.7.99
+      - nvidia-cuda-runtime-cu11==11.7.99
+      - nvidia-cudnn-cu11==8.5.0.96
+      - psutil==5.9.4
+      - pyparsing==3.0.9
+      - requests==2.28.2
+      - scikit-learn==1.2.1
+      - scipy==1.10.1
+      - threadpoolctl==3.1.0
+      - torch==1.13.1
+      - torch-geometric==2.2.0
+      - torch-scatter==2.1.0
+      - torch-sparse==0.6.16
+      - tqdm==4.64.1
+      - typing-extensions==4.5.0
+      - urllib3==1.26.14
+prefix: /home/ishan05/anaconda3/envs/graphs
