+++ torch_geometric/contrib/degree_quant/utils.py
@@ -1,0 +1,23 @@
+import torch
+import torch_scatter
+from torch_geometric.utils import degree
+
+
+def scatter_(name, src, index, dim=0, dim_size=None):
+    """
+    Method implemented to mask the minimum and maximum value in the tensor. 
+    """
+
+    assert name in ["add", "mean", "min", "max"]
+
+    op = getattr(torch_scatter, "scatter_{}".format(name))
+    out = op(src, index, dim, None, dim_size)
+    out = out[0] if isinstance(out, tuple) else out
+    if name == "max":
+        out[out < -10000] = 0
+    elif name == "min":
+        out[out > 10000] = 0
+    return out
+
+
+
