+++ torch_geometric/contrib/nn/conv/__init__.py
@@ -1,1 +1,6 @@
+from linear import * 
+from .messagepassing import *
+from GAT_conv import * 
+from GCN_conv import * 
+from GIN_conv import * 
 __all__ = classes = []
