--- torch_geometric/contrib/nn/models/degree_quant/__init__.py
@@ -1,8 +1,0 @@
-
-from .quantizer import *
-from .utils import *
-from .messagepassing import *
-from .GAT_conv import * 
-from .GCN_conv import * 
-from .GIN_conv import * 
-__all__ = classes = []
