--- torch_geometric/contrib/nn/models/degree_quant/experiment.py
@@ -1,7 +1,0 @@
-from data_handler import get_dataset
-from wrapper import GAT,GCN,GIN
-from training import cross_validation_with_val_set
-
-
-
-
